Replace all floats with doubles

+++ src/Nexus.Tests/MathUtilityTests.cs
@@ -8,22 +8,22 @@ namespace Nexus.Tests
 		[Test]
 		public void CanClampFloatValueInsideRange()
 		{
-			float result = MathUtility.Clamp(1.0f, 0.0f, 2.0f);
-			Assert.AreEqual(1.0f, result);
+			double result = MathUtility.Clamp(1.0, 0.0, 2.0);
+			Assert.AreEqual(1.0, result);
 		}
 
 		[Test]
 		public void CanClampFloatValueBelowRange()
 		{
-			float result = MathUtility.Clamp(-3.0f, 0.0f, 2.0f);
-			Assert.AreEqual(0.0f, result);
+			double result = MathUtility.Clamp(-3.0, 0.0, 2.0);
+			Assert.AreEqual(0.0, result);
 		}
 
 		[Test]
 		public void CanClampFloatValueAboveRange()
 		{
-			float result = MathUtility.Clamp(3.0f, 0.0f, 2.0f);
-			Assert.AreEqual(2.0f, result);
+			double result = MathUtility.Clamp(3.0, 0.0, 2.0);
+			Assert.AreEqual(2.0, result);
 		}
 
 		[Test]
