+++ src/Nexus.Tests/TestUtility.cs
@@ -4,9 +4,9 @@ namespace Nexus.Tests
 {
 	internal static class TestUtility
 	{
-		public static void AssertAreRoughlyEqual(float expectedValue, float actualValue)
+		public static void AssertAreRoughlyEqual(double expectedValue, double actualValue)
 		{
-			if (System.Math.Abs(expectedValue - actualValue) > 0.01f)
+			if (System.Math.Abs(expectedValue - actualValue) > 0.01)
 				Assert.Fail("Expected value: " + expectedValue + "; actual value: " + actualValue);
 		}
 	}
