+++ src/Nexus/MathUtility.cs
@@ -2,10 +2,10 @@ namespace Nexus
 {
 	public static class MathUtility
 	{
-		public const float PI = (float) System.Math.PI;
-		public const float PI_OVER_2 = (float) (System.Math.PI / 2.0f);
-		public const float PI_OVER_4 = (float) (System.Math.PI / 4.0f);
-		public const float TWO_PI = (float)(System.Math.PI * 2.0f);
+		public const double PI = (double) System.Math.PI;
+		public const double PI_OVER_2 = (double) (System.Math.PI / 2.0);
+		public const double PI_OVER_4 = (double) (System.Math.PI / 4.0);
+		public const double TWO_PI = (double)(System.Math.PI * 2.0);
 
 		/// <summary>
 		/// Clamps a value to an interval.
@@ -14,7 +14,7 @@ namespace Nexus
 		/// <param name="min">The lower clamp threshold.</param>
 		/// <param name="max">The upper clamp threshold.</param>
 		/// <returns>The clamped value.</returns>
-		public static float Clamp(float value, float min, float max)
+		public static double Clamp(double value, double min, double max)
 		{
 			value = (value > max) ? max : value;
 			value = (value < min) ? min : value;
@@ -49,12 +49,12 @@ namespace Nexus
 			return value;
 		}
 
-		public static int Floor(float value)
+		public static int Floor(double value)
 		{
 			return (int) System.Math.Floor(value);
 		}
 
-		public static int Ceiling(float value)
+		public static int Ceiling(double value)
 		{
 			return (int)System.Math.Ceiling(value);
 		}
@@ -98,7 +98,7 @@ namespace Nexus
 		/// <param name="value2">The upper interpolation bound.</param>
 		/// <param name="amount">The interpolation parameter.</param>
 		/// <returns>The interpolated value.</returns>
-		public static float Lerp(float value1, float value2, float amount)
+		public static double Lerp(double value1, double value2, double amount)
 		{
 			return (1 - amount) * value1 + amount * value2;
 			//return Lerp(value1, value2, 0, 1, amount); // This is the same thing
@@ -111,7 +111,7 @@ namespace Nexus
 		/// <param name="value2">The upper interpolation bound.</param>
 		/// <param name="amount">The interpolation parameter.</param>
 		/// <returns>The interpolated value.</returns>
-		public static int Lerp(int value1, int value2, float amount)
+		public static int Lerp(int value1, int value2, double amount)
 		{
 			return (int) ((1 - amount) * value1 + amount * value2);
 		}
@@ -123,31 +123,31 @@ namespace Nexus
 		/// <param name="value2">The upper interpolation bound.</param>
 		/// <param name="amount">The interpolation parameter.</param>
 		/// <returns>The interpolated value.</returns>
-		public static byte Lerp(byte value1, byte value2, float amount)
+		public static byte Lerp(byte value1, byte value2, double amount)
 		{
 			return (byte)((1 - amount) * value1 + amount * value2);
 		}
 
-		public static float Lerp(float value1, float value2, float startAmount, float endAmount, float amount)
+		public static double Lerp(double value1, double value2, double startAmount, double endAmount, double amount)
 		{
 			return (((value2 - value1) * amount) + ((value1 * endAmount) + (value2 * startAmount))) / (endAmount - startAmount);
 		}
 
-		public static float Log2(float d)
+		public static double Log2(double d)
 		{
-			return (float) System.Math.Log(d, 2.0f);
+			return (double) System.Math.Log(d, 2.0);
 		}
 
-		public static float PerspectiveInterpolate(float value1, float value2, float w1, float w2, float amountRangeStart, float amountRangeEnd, float amount)
+		public static double PerspectiveInterpolate(double value1, double value2, double w1, double w2, double amountRangeStart, double amountRangeEnd, double amount)
 		{
 			// Don't ask me how I got this... it's mostly from page 124 of 3D Game Engine Design,
 			// with a correction as noted on http://www.geometrictools.com/Books/GameEngineDesign2/BookCorrections.html.
-			float numerator = (((value2 * w1) - (value1 * w2)) * amount) + ((value1 * w2 * amountRangeEnd) - (value2 * w1 * amountRangeStart));
-			float denominator = ((w1 - w2) * amount) + ((w2 * amountRangeEnd) - (w1 * amountRangeStart));
+			double numerator = (((value2 * w1) - (value1 * w2)) * amount) + ((value1 * w2 * amountRangeEnd) - (value2 * w1 * amountRangeStart));
+			double denominator = ((w1 - w2) * amount) + ((w2 * amountRangeEnd) - (w1 * amountRangeStart));
 			return numerator / denominator;
 		}
 
-		public static int Round(float value)
+		public static int Round(double value)
 		{
 			return (int) System.Math.Round(value);
 		}
@@ -159,7 +159,7 @@ namespace Nexus
 		/// <param name="b">The upper threshold position.</param>
 		/// <param name="x">The input parameter.</param>
 		/// <returns>The interpolated value.</returns>
-		public static float SmoothStep(float a, float b, float x)
+		public static double SmoothStep(double a, double b, double x)
 		{
 			if (x < a)
 				return 0;
@@ -169,95 +169,95 @@ namespace Nexus
 			return x * x * (3 - 2 * x);
 		}
 
-		public static float Saturate(float value)
+		public static double Saturate(double value)
 		{
-			value = (value > 1.0f) ? 1.0f : value;
-			value = (value < 0.0f) ? 0.0f : value;
+			value = (value > 1.0) ? 1.0 : value;
+			value = (value < 0.0) ? 0.0 : value;
 			return value;
 		}
 
-		public static float Exp(float value)
+		public static double Exp(double value)
 		{
-			return (float)System.Math.Exp(value);
+			return (double)System.Math.Exp(value);
 		}
 
-		public static float Pow(float x, float y)
+		public static double Pow(double x, double y)
 		{
-			return (float)System.Math.Pow(x, y);
+			return (double)System.Math.Pow(x, y);
 		}
 
-		public static float Sqrt(float value)
+		public static double Sqrt(double value)
 		{
-			return (float) System.Math.Sqrt(value);
+			return (double) System.Math.Sqrt(value);
 		}
 
-		public static float ToDegrees(float radians)
+		public static double ToDegrees(double radians)
 		{
-			return (radians * 57.29578f);
+			return (radians * 57.29578);
 		}
 
-		public static float ToRadians(float degrees)
+		public static double ToRadians(double degrees)
 		{
-			return (degrees * 0.01745329f);
+			return (degrees * 0.01745329);
 		}
 
-		public static float Cos(float radians)
+		public static double Cos(double radians)
 		{
-			return (float) System.Math.Cos(radians);
+			return (double) System.Math.Cos(radians);
 		}
 
-		public static float Acos(float radians)
+		public static double Acos(double radians)
 		{
-			return (float)System.Math.Acos(radians);
+			return (double)System.Math.Acos(radians);
 		}
 
-		public static float Sin(float radians)
+		public static double Sin(double radians)
 		{
-			return (float) System.Math.Sin(radians);
+			return (double) System.Math.Sin(radians);
 		}
 
-		public static float Asin(float sin)
+		public static double Asin(double sin)
 		{
-			return (float) System.Math.Asin(sin);
+			return (double) System.Math.Asin(sin);
 		}
 
-		public static float Tan(float radians)
+		public static double Tan(double radians)
 		{
-			return (float) System.Math.Tan(radians);
+			return (double) System.Math.Tan(radians);
 		}
 
-		public static float Atan2(float y, float x)
+		public static double Atan2(double y, double x)
 		{
-			return (float)System.Math.Atan2(y, x);
+			return (double)System.Math.Atan2(y, x);
 		}
 
-		public static void Swap(ref float v1, ref float v2)
+		public static void Swap(ref double v1, ref double v2)
 		{
-			float temp = v1;
+			double temp = v1;
 			v1 = v2;
 			v2 = temp;
 		}
 
-		public static bool IsZero(float value)
+		public static bool IsZero(double value)
 		{
-			return (System.Math.Abs(value) < 1.0E-6f);
+			return (System.Math.Abs(value) < 1.0E-6);
 		}
 
-		public static bool Quadratic(float A, float B, float C, out float t0, out float t1)
+		public static bool Quadratic(double A, double B, double C, out double t0, out double t1)
 		{
 			// Find quadratic discriminant
-			float discrim = B * B - 4.0f * A * C;
-			if (discrim < 0.0f)
+			double discrim = B * B - 4.0 * A * C;
+			if (discrim < 0.0)
 			{
-				t0 = t1 = float.MinValue;
+				t0 = t1 = double.MinValue;
 				return false;
 			}
-			float rootDiscrim = Sqrt(discrim);
+			double rootDiscrim = Sqrt(discrim);
 
 			// Compute quadratic _t_ values
-			float q;
-			if (B < 0) q = -0.5f * (B - rootDiscrim);
-			else q = -0.5f * (B + rootDiscrim);
+			double q;
+			if (B < 0) q = -0.5 * (B - rootDiscrim);
+			else q = -0.5 * (B + rootDiscrim);
 			t0 = q / A;
 			t1 = C / q;
 			if (t0 > t1) Swap(ref t0, ref t1);
